+++ jadx-core/build.gradle.kts
@@ -7,14 +7,6 @@ dependencies {
 
 	implementation("com.google.code.gson:gson:2.10.1")
 
-	// TODO: move resources decoding to separate plugin module
-	implementation("com.android.tools.build:aapt2-proto:8.3.2-10880808")
-	implementation("com.google.protobuf:protobuf-java") {
-		version {
-			require("3.25.3") // version 4 conflict with `aapt2-proto`
-		}
-	}
-
 	testImplementation("org.apache.commons:commons-lang3:3.14.0")
 
 	testImplementation(project(":jadx-plugins:jadx-dex-input"))
