--- jadx-core/src/main/java/jadx/core/xmlgen/IResParser.java
@@ -1,13 +1,0 @@
-package jadx.core.xmlgen;
-
-import java.io.IOException;
-import java.io.InputStream;
-
-public interface IResParser {
-
-	void decode(InputStream inputStream) throws IOException;
-
-	ResourceStorage getResStorage();
-
-	BinaryXMLStrings getStrings();
-}
