+++ jadx-core/src/main/java/jadx/core/xmlgen/IResTableParser.java
@@ -1,0 +1,15 @@
+package jadx.core.xmlgen;
+
+import java.io.IOException;
+import java.io.InputStream;
+
+public interface IResTableParser {
+
+	void decode(InputStream inputStream) throws IOException;
+
+	ResContainer decodeFiles(InputStream inputStream) throws IOException;
+
+	ResourceStorage getResStorage();
+
+	BinaryXMLStrings getStrings();
+}
