+++ jadx-core/src/main/java/jadx/core/xmlgen/ResDecoder.java
@@ -5,25 +5,28 @@ import java.io.InputStream;
 
 import jadx.api.ResourceFile;
 import jadx.api.ResourceType;
+import jadx.api.ResourcesLoader;
+import jadx.api.plugins.resources.IResTableParserProvider;
 import jadx.core.dex.nodes.RootNode;
 import jadx.core.utils.exceptions.JadxRuntimeException;
 
 public class ResDecoder {
 
-	public static IResParser decode(RootNode root, ResourceFile resFile, InputStream is) throws IOException {
+	public static IResTableParser decode(RootNode root, ResourceFile resFile, InputStream is) throws IOException {
 		if (resFile.getType() != ResourceType.ARSC) {
 			throw new IllegalArgumentException("Unexpected resource type for decode: " + resFile.getType() + ", expect ARSC");
 		}
-		IResParser parser = null;
-		String fileName = resFile.getOriginalName();
-		if (fileName.endsWith(".arsc")) {
-			parser = new ResTableParser(root);
+
+		IResTableParser parser = null;
+		for (IResTableParserProvider provider : ResourcesLoader.getResTableParserProviders()) {
+			parser = provider.getInstance(root, resFile);
+			if (parser != null) {
+				break;
+			}
 		}
-		if (fileName.endsWith(".pb")) {
-			parser = new ResProtoParser(root);
-		}
+
 		if (parser == null) {
-			throw new JadxRuntimeException("Unknown type of resource file: " + fileName);
+			throw new JadxRuntimeException("Unknown type of resource file: " + resFile.getOriginalName());
 		}
 		parser.decode(is);
 		return parser;
