+++ jadx-plugins/jadx-res-proto-input/src/main/java/jadx/plugins/input/res/proto/ResContainerProtoXmlFactory.java
@@ -1,0 +1,33 @@
+package jadx.plugins.input.res.proto;
+
+import jadx.api.ICodeInfo;
+import jadx.api.ResourceFile;
+import jadx.api.ResourceType;
+import jadx.api.plugins.resources.IResContainerFactory;
+import jadx.core.dex.nodes.RootNode;
+import jadx.core.xmlgen.ResContainer;
+import org.jetbrains.annotations.Nullable;
+
+import java.io.IOException;
+import java.io.InputStream;
+
+public class ResContainerProtoXmlFactory implements IResContainerFactory {
+	private ResXmlProtoParser xmlParser;
+
+	@Override
+	@Nullable
+	public synchronized ResContainer create(RootNode root, ResourceFile resFile, InputStream inputStream) throws IOException {
+		if (!root.isProto() || resFile.getType() != ResourceType.XML) {
+			return null;
+		}
+		ICodeInfo content = getInstance(root).parse(inputStream);
+		return ResContainer.textResource(resFile.getDeobfName(), content);
+	}
+
+	synchronized private ResXmlProtoParser getInstance(RootNode root) {
+		if (xmlParser == null) {
+			xmlParser = new ResXmlProtoParser(root);
+		}
+		return xmlParser;
+	}
+}
