+++ jadx-plugins/jadx-res-proto-input/src/main/resources/META-INF/services/jadx.api.plugins.JadxPlugin
@@ -1,0 +1,1 @@
+jadx.plugins.input.res.proto.ResProtoPlugin
