+++ settings.gradle.kts
@@ -19,6 +19,7 @@ include("jadx-plugins:jadx-java-convert")
 include("jadx-plugins:jadx-rename-mappings")
 include("jadx-plugins:jadx-kotlin-metadata")
 include("jadx-plugins:jadx-xapk-input")
+include("jadx-plugins:jadx-res-proto-input")
 
 include("jadx-plugins:jadx-script:jadx-script-plugin")
 include("jadx-plugins:jadx-script:jadx-script-runtime")
